% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_status.R
\name{assign_status}
\alias{assign_status}
\title{Assign native/introduced status to observations}
\usage{
assign_status(observations, sp_range)
}
\arguments{
\item{observations}{A list containing observation data and metadata.}

\item{sp_range}{A Distribution object (typicaly created via \code{\link{get_range}}).}
}
\value{
Spatial points with an attribute `status` which is either `native`, `introduced`, or `unknown`.
}
\description{
This function intersects observation locations and documented species range
to assign native/introduced status to each data point.
}
\details{
This function performs a spatial intersection of native range and introduced range (as spatial polygon features) with the point locations of observations.
}
