% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_response.R
\name{check_response}
\alias{check_response}
\title{Check response data}
\usage{
check_response(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame matching the required format.
}
\description{
Check if the input data frame matches the required format. If the input is
not a data frame, it will be replaced with an empty data frame. Missing
columns are added and filled with NA.
}
