% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{config}
\alias{config}
\title{Default configuration for GBIFhandleR package}
\format{
A list with the following elements:
\describe{
  \item{COLUMNS}{Character vector of default column names for GBIF data.}
  \item{GBIFMAX}{Numeric value specifying the maximum number of records to retrieve from GBIF (default: 100).}
}
}
\usage{
config
}
\description{
A list containing default configuration values used across the package.
}
\keyword{datasets}
