% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_records.R
\name{count_records}
\alias{count_records}
\title{Count records for a species in GBIF}
\usage{
count_records(
  species_name,
  hasCoordinate = TRUE,
  hasGeospatialIssue = FALSE,
  ...
)
}
\arguments{
\item{species_name}{A character string specifying the scientific name of
the species to search for.}

\item{hasCoordinate}{Logical. If `TRUE`, only count records with coordinate
data. The default is `TRUE`.}

\item{hasGeospatialIssue}{Logical. If `FALSE`, excludes records with known
geospatial issues. Default is `FALSE`.}

\item{...}{Additional parameters to pass to `rgbif::occ_search()` for
further filtering (e.g., year, country).}
}
\value{
An integer representing the observation count.
}
\description{
This function queries the GBIF database to count the number of observation
records for a specified species without retrieving the dataset.
}
\details{
- The function uses the `rgbif::occ_search()` function with a `limit` of
    `0` to efficiently count records without downloading data.
- Filters such as `hasCoordinate` and `hasGeospatialIssue` allow for
    restricting the count to specific types of records.
}
\examples{
# Example 1: Count records for a species with coordinates and no geospatial issues
count_records("Ficaria verna Huds.")

# Example 2: Count records for a species in a specific year
count_records("Ficaria verna Huds.", year = "2020")

}
