% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{Search GBIF for observations of a species}
\usage{
get_observations(
  species_name,
  match_name = TRUE,
  start_year = 1950,
  end_year = format(Sys.Date(), "\%Y"),
  return_full = TRUE
)
}
\arguments{
\item{species_name}{A character string specifying the scientific name
of the species to search for.}

\item{match_name}{Logical. If `TRUE`, attempts to match the species name
with a GBIF taxon using a helper function (`match_gbif_name`).
The default is `TRUE`.}

\item{start_year}{Numeric. The starting year for observations to include.
The default is `1950`.}

\item{end_year}{Numeric. The ending year for observations to include.
The default is the current year.}

\item{return_full}{Logical. If `TRUE`, returns a list containing full
details of the search, including grid information and the matched GBIF name.
If `FALSE`, the function only returns a data frame of observations.
The default is `TRUE`.}
}
\value{
A list or data frame:
  - If `return_full = TRUE`, returns a list with the following elements:
      - `data`: A data frame of observations with requested columns.
      - `grid`: A data frame of the grid used for dividing large datasets.
      - `inputname`: The input species name.
      - `gbifname`: The matched GBIF species name.
  - If `return_full = FALSE`, returns only the data frame of observations.
}
\description{
This function queries the GBIF database for observation records of a
specified species within a given year range, optionally handling name
matching and data limits imposed by the GBIF API.
It allows retrieval of either a simple data frame or a list containing
additional information related to the retrieval of the data.
}
\details{
- The function uses the `rgbif` package to interact with GBIF's API.
- Observations are restricted to those with coordinates and without
geospatial issues.
- For large datasets (exceeding 100,000 records), the function divides the
search area into a grid and downloads data in smaller chunks.
}
\examples{
# Example 1: Basic search for a species
observations <- get_observations("Grewia rogersii")

# Example 2: Retrieve only a simplified data frame
\dontrun{
observations <- get_observations("Tragopogon dubius", return_full = FALSE)
}

}
