% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations_quietly.R
\name{get_observations_quietly}
\alias{get_observations_quietly}
\title{Search GBIF for observations of a species without notifications and warnings}
\usage{
get_observations_quietly(...)
}
\arguments{
\item{...}{Any additional parameter that can be passed to `rgbif::occ_search`.}
}
\value{
The output of \code{\link{get_observations}}.
}
\description{
A wrapper around \code{\link{get_observations}} that suppresses all messages,
warnings, and output during execution.
}
