% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmass.R
\docType{data}
\name{landmass}
\alias{landmass}
\title{Natural Earth: Landmass at scale 1:110}
\format{
A `sf::sfc_POLYGON` with the following attributes:
\describe{
  \item{geometry}{Polygon geometries representing landmasses.}
}
}
\source{
Natural Earth: https://www.naturalearthdata.com/
}
\usage{
landmass
}
\description{
An `sf::sfc_POLYGON` object representing global landmasses. This dataset is
used as a default background map in the package's plotting functions.
}
\examples{
plot(landmass)

}
\keyword{datasets}
