% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_observations.R
\name{load_observations}
\alias{load_observations}
\title{Read species observations}
\usage{
load_observations(filename)
}
\arguments{
\item{filename}{Character. Input file.}
}
\value{
A list containing observation data and metadata.
}
\description{
Attempt to read species observations and metadata exported via \code{\link{save_observations}}.
}
\details{
This function attempts to read species observation data from a .csv file along with the original query name and the matched GBIF name of the species.
In case valid metadata cannot be found in the input file, the first two words of the file name are extracted and set as input file name.
}
