% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gbif_name.R
\name{match_gbif_name}
\alias{match_gbif_name}
\title{Match an input species name with GBIF}
\usage{
match_gbif_name(species_name, rank = "SPECIES")
}
\arguments{
\item{species_name}{A character string specifying the name of the species
to match. This can be a scientific name or common name.}

\item{rank}{A character string specifying the taxonomic rank to match.
The default is `"SPECIES"`. Other options include `"GENUS"`,
`"FAMILY"`, etc.}
}
\value{
A character string containing the matched scientific name from GBIF.
  If no match is found, returns `NA`.
}
\description{
This function attempts to match a given species name with the GBIF
taxonomic backbone to find its standardized scientific name.
}
\details{
- The function uses the `rgbif::name_backbone()` function to query GBIF’s
    taxonomic backbone.
- If no match is found, the function returns `NA` and does not raise an
    error.
- Matching is not strict, allowing for flexible name matching.
- The rank parameter can be used to refine the matching process (e.g.,
    only match species or genera).
}
\examples{
# Example 1: Match a species name. The following will perform fuzzy matching and return the corrected, full species name: "Pterocarpus rotundifolius (Sond.) Druce"
match_gbif_name("Pterocarpus rotundifolia")

# Example 2: Match a genus name
match_gbif_name("Panthera", rank = "GENUS")

}
