% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution-methods.R
\name{plot,Distribution,ANY-method}
\alias{plot,Distribution,ANY-method}
\title{Plot a Distribution object}
\usage{
\S4method{plot}{Distribution,ANY}(x, native_col = "aquamarine4", introduced_col = "firebrick4", ...)
}
\arguments{
\item{x}{An object of class \code{Distribution}, containing the species name and its geographical distribution.}

\item{native_col}{A character string specifying the coluor used to represent the native range. Defaults to \code{"aquamarine4"}.}

\item{introduced_col}{A character string specifying the colour used to represent the introduced range. Defaults to \code{"firebrick4"}.}

\item{...}{Additional graphical parameters to pass to the underlying \code{\link[graphics]{plot}} function.}
}
\description{
Visualizes the geographical distribution of a species represented by a \code{Distribution} object.
The native and introduced ranges are overlaid on the TDWG regions map with customisable colours.
}
\details{
This method plots the TDWG regions map and overlays the native and introduced ranges of the species.
The native regions are displayed in \code{native_col}, while the introduced regions are displayed in \code{introduced_col}.
The TDWG regions are provided by the preloaded \code{tdwg} dataset.
}
\examples{
\dontrun{
# Example: Plotting the distribution of Quercus robur
qr_dist <- get_range("Quercus robur")
plot(qr_dist, native_col = "green", introduced_col = "red")
}

}
\seealso{
\link{get_range}, \link{Distribution-class}
}
