% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_observations.R
\name{save_observations}
\alias{save_observations}
\title{Save species observations as comma separated file}
\usage{
save_observations(observations, filename)
}
\arguments{
\item{observations}{A list containing observation data and metadata as returned by \code{\link{get_observations}}.}

\item{filename}{Character. Output file.}
}
\description{
This function stores the output of \code{\link{get_observations}} as a .csv file
along with metadata.
}
\details{
This function stores the observation data along with the original query name and the matched GBIF name of the species.
}
