% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_powo_code_by_name.R
\name{search_powo_code_by_name}
\alias{search_powo_code_by_name}
\title{Search POWO code by species name}
\usage{
search_powo_code_by_name(species_name, preview = FALSE)
}
\arguments{
\item{species_name}{A character string representing the scientific name of the species to search for.}

\item{preview}{A logical value indicating whether to preview the first few results. If `TRUE`, prints a preview of up to six results. Defaults to `FALSE`.}
}
\value{
A character string containing the POWO URL for the first match if found; otherwise, `NULL`.
}
\description{
This function searches the Plants of the World Online (POWO) database for a
species by name and returns the unique POWO code (URL) for the first match.
}
\details{
The function sends a GET request to the POWO API to search for the specified species.
If results are found, the unique POWO code (URL) of the first match is returned.
If `preview = TRUE`, the function prints a summary of the first few results.
}
\examples{
\dontrun{
# Search for the POWO code of "Quercus robur"
powo_code <- search_powo_code_by_name("Quercus robur")
print(powo_code)

# Preview search results
search_powo_code_by_name("Quercus robur", preview = TRUE)
}

}
