% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_grid.R
\name{split_grid}
\alias{split_grid}
\title{Split grid cells with many observations into four quarters}
\usage{
split_grid(grid, min_collapse_lat = 1, min_collapse_lon = 1)
}
\arguments{
\item{grid}{The initial grid (a data.frame with the columns `latmin`,
`lonmin`, `latmax`, and `lonmax`, as well as a column `num_records`
which indicates the current number of records found in the grid cells).}

\item{min_collapse_lat}{Numeric. A threshold providing the minimum
height of a grid cell in degrees latitude. The default is 1.}

\item{min_collapse_lon}{Numeric. A threshold providing the minimum
width of a grid cell in degrees longitude The default is 1.}
}
\value{
A data.frame containing a new grid where cells with too many
    observations have been split into quarters by longitude and latitude.
}
\description{
Split grid into four quarters if the number of observations
within the grid cell is larger than the maximum allowed by
GBIF. Grid cells are only split as long as they retain the
minimum height and width as provided through min_collapse_lat
and min_collapse_lon. If the minimum height or width criterium
is no longer met after splitting, the grid cell is returned as
is and the number of observations is set to the maximum value
of 100'000. Thus, there will only 100'000 observations be
extracted per minimum grid cell during the analyses.
}
