% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwg.R
\docType{data}
\name{tdwg}
\alias{tdwg}
\title{Biodiversity Information Standards: World Geographical Scheme for Recording Plant Distributions}
\format{
A `sf::st_sfc` with the following attributes:
\describe{
  \item{code}{TDWG location code.}
  \item{name}{Location name.}
  \item{level}{Locality level.}
  \item{geometry}{Polygon geometries.}
}
}
\source{
Biodiversity Information Standards (TDWG): https://github.com/tdwg/wgsrpd/tree/master/geojson
}
\usage{
tdwg
}
\description{
A simple feature collection delineating locations following the World Geographical Scheme for Recording Plant Distributions (WGSRPD).
}
\examples{
plot(tdwg)

}
\keyword{datasets}
